separated fury and plotly into two files; added viz class to viz which chooses between them
